add new LambdaRenderer Class
The lambda renderer utilizes the serverless method to do it's page rendering. you can pass all of the existing arguments to the `render()` method now to the lambda renderer in Lambda, allowing for full control just like in the PhantomJSRenderer class.

Todo: need to repackage phantom-js so I can push it into lambda and test everything. Happy with the unit tests but need to double check the lambda handler itself